backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
supabase
resend
youtube-transcript-api
pydantic
python-dotenv
httpx
google-api-python-client
bcrypt
